[REFACTOR(LottoTest,LottoResultTest)] 패키지 수정
